+++ src/features/friends/ui/friend-list.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 
-import useFriends from '../model/use-friends'
+import { useFriends } from '../model'
 
-import FriendItem from './friend-item'
+import { FriendItem } from './index'
 
 import type { Friend } from '@/entities/friends/models/friend.types'
